Correction of many anoying warnings
git-svn-id: https://stlport.svn.sourceforge.net/svnroot/stlport/branches/R451_dev/STLport@313 01b27279-081b-0410-8cf9-d9d9e0fb0389

+++ test/regression/copy2.cpp
@@ -17,7 +17,7 @@ int copy2_test(int, char**)
 {
   cout<<"Results of copy2_test:"<<endl;
   vector <int> v(10);
-  for(int i = 0; i < v.size(); i++)
+  for(size_t i = 0; i < v.size(); i++)
     v[i] = i;
   ostream_iterator<int> iter(cout, " ");
   copy(v.begin(), v.end(), iter);
